worker/requirements.txt
redis==5.2.1
psycopg2-binary==2.9.10
